# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.AuditLogRequest:
    properties:
      cluster_context:
        type: string
      execution_result:
        additionalProperties: true
        type: object
      execution_status:
        enum:
        - success
        - failed
        - cancelled
        type: string
      generated_command:
        type: string
      ip_address:
        type: string
      namespace_context:
        type: string
      query_text:
        type: string
      safety_level:
        enum:
        - safe
        - warning
        - dangerous
        type: string
      session_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    required:
    - execution_status
    - generated_command
    - query_text
    - safety_level
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@kubechat.dev
    name: KubeChat Support Team
    url: https://github.com/pramodksahoo/kubechat
  description: |-
    Kubernetes-native AI platform with multi-provider support for enterprise deployments

    This API provides comprehensive management capabilities for:
    - Kubernetes cluster operations and monitoring
    - Multi-provider AI/ML services (OpenAI, Anthropic, Google, Ollama)
    - Real-time WebSocket communication
    - Advanced cost tracking and budget management
    - Enterprise security and compliance features
    - Comprehensive audit logging and monitoring
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: KubeChat Enterprise Platform API
  version: 1.0.0
paths:
...
  /audit/dangerous:
    get:
      description: Retrieves audit logs for operations classified as dangerous
      parameters:
      - description: Start time (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end_time
        type: string
      - default: 50
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dangerous operations retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve dangerous operations
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get dangerous operations
      tags:
      - Audit & Compliance
  /audit/failed:
    get:
      description: Retrieves audit logs for operations that failed during execution
      parameters:
      - description: Start time (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end_time
        type: string
      - default: 50
        description: Limit number of results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Failed operations retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve failed operations
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get failed operations
      tags:
      - Audit & Compliance
...
  /audit/logs:
    get:
      description: Retrieves audit logs with optional filtering by user, session,
        safety level, status, and time range
      parameters:
      - description: User ID to filter by
        in: query
        name: user_id
        type: string
      - description: Session ID to filter by
        in: query
        name: session_id
        type: string
      - description: Safety level (safe, warning, dangerous)
        in: query
        name: safety_level
        type: string
      - description: Execution status (success, failed, cancelled)
        in: query
        name: status
        type: string
      - description: Start time (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339 format)
        in: query
        name: end_time
        type: string
      - default: 50
        description: Limit number of results (1-1000)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit logs retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid query parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve audit logs
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get audit logs with filtering
      tags:
      - Audit & Compliance
    post:
      consumes:
      - application/json
      description: Creates a new audit log entry for tracking user actions
      parameters:
      - description: Audit log data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AuditLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Audit log created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create audit log
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new audit log entry
      tags:
      - Audit & Compliance
  /audit/logs/{id}:
    get:
      description: Retrieves a specific audit log entry by its ID
      parameters:
      - description: Audit log ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Audit log retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid audit log ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Audit log not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log by ID
      tags:
      - Audit & Compliance
...
  /audit/summary:
    get:
      description: Returns summary statistics and analytics for audit logs within
        a time range
      parameters:
      - description: Start time for summary (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time for summary (RFC3339 format)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Audit log summary retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get audit log summary
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get audit log summary statistics
      tags:
      - Audit & Compliance
  /audit/verify-integrity:
    post:
      consumes:
      - application/json
      description: Verifies the integrity of audit logs within a specified ID range
      parameters:
      - description: Integrity verification request
        in: body
        name: request
        required: true
        schema:
          properties:
            end_id:
              type: integer
            start_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Integrity verification completed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request format or ID range
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Integrity check failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Verify audit log integrity
      tags:
      - Audit & Compliance
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for service-to-service communication
    in: header
    name: X-API-Key
    type: apiKey
    x-extension-openapi: '{"example": "value"}'
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"